Abandon the OBS build tool and use mock instead
The OBS build tool has depdency resolution limitations that make it
unable to build some packages.  In particular it only recognizes a
single verison of a given package and so is *quite* unable to handle
our multi-package-version repos.

Since we are using mock for all RPM builds, we might as well just use
Fedora as the build platform to get much more recent versions of mock.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>